Move private recommendation methods to base class
